+++ src/app/post/index.module.scss
@@ -38,135 +38,3 @@
     flex-direction: column;
   }
 }
-
-.postImage {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  width: 25rem;
-
-  & > p {
-    width: 25rem;
-    padding-right: 0.5rem;
-    text-align: right;
-    cursor: pointer;
-    font-size: 12px;
-    & > label {
-      color: var(--main-color);
-      border-bottom: 1px solid var(--main-color);
-    }
-
-    @include desktop {
-      width: 35rem;
-    }
-
-    @include mobile {
-      width: 50rem;
-    }
-  }
-
-  @include desktop {
-    width: 35rem;
-  }
-
-  @include mobile {
-    width: 50rem;
-  }
-}
-
-.uploadProductImage {
-  position: relative;
-  width: 25rem;
-  height: 25rem;
-  background-color: #ffffff;
-  border: 1px solid #ccc;
-  border-radius: 5px;
-  margin: 10px 0;
-
-  & > input {
-    display: none;
-  }
-
-  @include desktop {
-    width: 35rem;
-    height: 35rem;
-  }
-
-  @include mobile {
-    width: 50rem;
-    height: 50rem;
-  }
-}
-
-.postProductInfo {
-  width: 30rem;
-  display: flex;
-  flex-direction: column;
-  align-items:baseline;
-
-  @include desktop {
-    width: 35rem;
-  }
-
-  @include mobile {
-    width: 50rem;
-    margin-top: 5rem;
-  }
-}
-
-.inputBox {
-  width: 100%;
-  position: relative;
-
-  .detailTextarea {
-    height: 20rem;
-    padding-top: 10px;
-  }
-
-  & > input, textarea {
-    width: 100%;
-    height: 5.5rem;
-    border-radius: 5px;
-    background-color: #ffffff;
-    margin: 10px 0;
-    border: 1px solid #ccc;
-    outline: none;
-    padding-left: 10px;
-    padding-right: 45px;
-    font-size: 1.6rem;
-    &:focus {
-      border-color: #808080;
-      box-shadow: 0 0 5px rgba(128, 128, 128, 0.5);
-    }
-    ::placeholder {
-      color: transparent;
-    }
-  }
-}
-
-.icon {
-  position: absolute;
-  width: 15px;
-  height: 15px;
-  top: calc(50% - 7.5px);
-  right: 15px;
-  color: gray;
-}
-
-.productDetailResetIcon {
-  top: 25px;
-}
-
-.submitButton {
-  background-color: var(--main-color);
-  color: #ffffff;
-  height: 3.5rem;
-  width: 100%;
-  margin-top: 3rem;
-  font-size: 2rem;
-  font-weight: bold;
-  letter-spacing: 0.5rem;
-  border: none;
-  border-radius: 5px;
-  cursor: pointer;
-}
