+++ src/components/postPage/InputBox/index.module.scss
@@ -1,0 +1,44 @@
+@import '@/src/styles/media.scss';
+
+.inputBox {
+  width: 100%;
+  position: relative;
+
+  & > input, textarea {
+    width: 100%;
+    height: 5.5rem;
+    border-radius: 5px;
+    background-color: #ffffff;
+    margin: 10px 0;
+    border: 1px solid #ccc;
+    outline: none;
+    padding-left: 10px;
+    padding-right: 45px;
+    font-size: 1.6rem;
+    &:focus {
+      border-color: #808080;
+      box-shadow: 0 0 5px rgba(128, 128, 128, 0.5);
+    }
+    ::placeholder {
+      color: transparent;
+    }
+  }
+
+  & > textarea {
+    height: 20rem;
+    padding-top: 10px;
+  }
+}
+
+.icon {
+  position: absolute;
+  width: 15px;
+  height: 15px;
+  top: calc(50% - 7.5px);
+  right: 15px;
+  color: gray;
+}
+
+.productDetailResetIcon {
+  top: 25px;
+}
